+++ components/pine-autocomplete.ts
@@ -120,6 +120,17 @@ function getPineCompletions(
       }
     });
 
+    hints.update?.forEach((hint, index) => {
+      if (hint.column.toLowerCase().includes(word.toLowerCase()) || word === '') {
+        completions.push({
+          expression: hint.column,
+          section: 'Columns',
+          label: hint.column,
+          apply: `${hint.column} = `,
+        });
+      }
+    });
+
   }
 
   if (completions.length === 0) {